compiling execute

+++ design/EverythingEverywhereAllAtOnce/rtl/core/EXE/FunctionalUnits/ret_imm_op.sv
@@ -1,6 +1,6 @@
 module ret_imm_op(
-    input uint32_t imm64,
-    input uint32_t stack_ptr,
+    input uint64_t imm64,
+    input uint64_t stack_ptr,
     
     //eip is sent to branch
     output uint64_t sr_o
